+++ scripts/test/main.py
@@ -1,48 +1,13 @@
 #!/usr/bin/env python3
 
 import argparse
-import os
-import json
-import re
 import logging
-from api import fetch_course_info, BASE_URL
-#from validation import validate_class_name
-from dotenv import load_dotenv
+import os
+from api import get_course_information
 
-# Load environment variables
-load_dotenv()
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
-
-
-#Fetch course information from the Berkeley API.
-def get_course_information(term_id, class_name):
-    """
-    Fetch course information from the Berkeley API.
-    Args:
-        term_id (int): The term ID.
-        class_name (str): The class name.
-    Returns: str: JSON string with extracted information or error message.
-    """
-
-    """#Validate the class name and extract the subject area code and catalog number
-    Maybe won't need validaiton since we implemented try in api.py
-    validation_result, error = validate_class_name(class_name)
-    if error:
-        # If there's an error in validation, return the error message as JSON
-        return json.dumps(error, indent=4)
-    # Unpack the validation result into subject area code and catalog number
-    subject_area_code, catalog_number = validation_result
-    # Fetch the course information from the API
-    result = fetch_course_info(term_id, f"{subject_area_code}{catalog_number}")"""
-    
-        
-    # Fetch the course information from the API
-    result = fetch_course_info(term_id, class_name)
-
-    # Return the result as a JSON string
-    return json.dumps(result, indent=4)
 
 
 def main():
@@ -50,8 +15,8 @@ def main():
     Main function to handle command-line arguments and fetch course information.
     """
     parser = argparse.ArgumentParser(description="Fetch course information based on term ID and class name.")
-    parser.add_argument('term_id', type=int, nargs='?', default=2232, help="The term ID (e.g., '2232').")
-    parser.add_argument('class_names', type=str, nargs='*', help="One or more class names (e.g., 'data8', 'compsci189').")
+    parser.add_argument('term_id', type=int, nargs='?', default=2232, help="The term ID (e.g., 2232).")
+    parser.add_argument('class_names', type=str, nargs='*', help="One or more class names (e.g., data8, compsci189).")
 
     args = parser.parse_args()
